Add padding to page and migrate from shared
