+++ webui/react/src/components/Page.module.scss
@@ -1,0 +1,44 @@
+.base {
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+  height: inherit;
+  justify-content: flex-start;
+  overflow: auto;
+  width: 100%;
+
+  .body {
+    flex-shrink: 0;
+    height: max-content;
+    min-height: 0;
+    padding: 16px;
+    width: 100%;
+  }
+  &.bodyNoPadding {
+    .body {
+      padding: 0;
+      padding-bottom: 16px;
+    }
+  }
+  &.stickyHeader {
+    overflow: inherit;
+
+    .body {
+      overflow: auto;
+      padding-bottom: 0;
+    }
+  }
+  &.noScroll {
+    max-height: 100vh;
+
+    .body {
+      flex: 1 1 auto;
+    }
+  }
+}
+
+@media only screen and (max-width: $breakpoint-mobile) {
+  .base.noScroll {
+    max-height: inherit;
+  }
+}
