+++ webui/react/src/components/PageHeader/PageHeader.tsx
@@ -2,7 +2,7 @@ import React, { useMemo } from 'react';
 
 import Breadcrumb from 'components/kit/Breadcrumb';
 import Tooltip from 'components/kit/Tooltip';
-import { BreadCrumbRoute } from 'shared/components/Page';
+import { BreadCrumbRoute } from 'components/Page';
 import { CommonProps } from 'shared/types';
 
 import Link from '../Link';
