+++ webui/react/src/pages/Dashboard.module.scss
@@ -1,6 +1,4 @@
 .table {
-  padding-bottom: 16px;
-
   td {
     white-space: nowrap;
   }
