+++ webui/react/src/pages/ProjectDetails.module.scss
@@ -1,6 +1,5 @@
 .base {
   flex: 1 1 auto;
-  flex-grow: 1;
   height: 100%;
 }
 .tabPane {
