--- webui/react/src/shared/components/Page/Page.module.scss
@@ -1,31 +1,0 @@
-.base {
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  flex-grow: 1;
-  height: inherit;
-  justify-content: flex-start;
-  overflow: auto;
-  width: 100%;
-
-  .body {
-    height: max-content;
-    height: 100%;
-    min-height: 0;
-    padding: 16px;
-    width: 100%;
-  }
-}
-.bodyNoPadding {
-  .body {
-    padding: 0;
-  }
-}
-.stickyHeader {
-  overflow: inherit;
-
-  .body {
-    overflow: auto;
-    padding-bottom: 0;
-  }
-}
