--- webui/react/src/shared/components/Page/index.ts
@@ -1,3 +1,0 @@
-export * from './Page';
-import Page from './Page';
-export default Page;
